// src/user.services/user.service.interface.ts
export interface AuthenticateReturn {
  username: string;
  wallet: WalletData;
  safeProperties: SafeProperties;
  playerStats: PlayerStats;
  energy?: { energy: number, timeUntilNextRecharge: number | null };
  uuid: string;
  refreshToken: string;
  accessToken: string;
  message: string;
  success: string;
}
// ...
export interface PlayerStats {
    level: number;
    playerExp: number;
    availStatPoints: number;
    rank: string;
    statPointsSaved: {
      mainVocalist: number;
      rapper: number;
      leadDancer: number;
      leadVocalist: number;
      mainDancer: number;
      visual: number;
      leadRapper: number;
    };
  }
// ...
export interface SafeProperties {
    signupDate: number;
  }
// ...
export interface WalletData {
    smartWalletAddress: string;
    beatsBalance: string;
    nativeBalance: string;
    apeBalance: string;
  }
// ...
export interface User {
  deviceId: string
  userName: string
  password: string | ""
  email?: string
}


export interface PasskeyUser extends User {
  id: string
  publicKey: Uint8Array
  counter: number
}
// ...
export interface TokenScheme {
  refreshToken: string
  accessToken: string
  userName: string
}
// ...
interface PlayerLevel {
  kind: string;
  level: number;
  minExperiencePoints: string;
  maxExperiencePoints: string;
}

/**
 * Represents the profile settings of a player.
 *
 * @interface ProfileSettings
 * @property {string} kind - The type of profile settings.
 * @property {boolean} profileVisible - Indicates if the player's profile is visible.
 * @property {string} friendsListVisibility - The visibility of the friends list.
 */
interface ProfileSettings {
  kind: string;
  profileVisible: boolean;
  friendsListVisibility: string;
}
// ...
interface PlayerExperienceInfo {
  kind: string;
  currentExperiencePoints: string;
  currentLevel: PlayerLevel;
  nextLevel: PlayerLevel;
}
// ...
export interface PlayerInfo {
  kind: string;
  playerId: string;
  displayName: string;
  avatarImageUrl: string;
  bannerUrlPortrait: string;
  bannerUrlLandscape: string;
  profileSettings: ProfileSettings;
  experienceInfo: PlayerExperienceInfo;
  title: string;
}